% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASM_PrePostProcessorFunctions.R
\name{CASMNodeTablePreparer}
\alias{CASMNodeTablePreparer}
\title{A function to create a CASM node location table given an REC network with additional attributes of CASM tributary labels and CASM tributary distances}
\usage{
CASMNodeTablePreparer(
  CASMRECNetwork = RECReachNetwork,
  NetworkLabelList = NetworkLabelList,
  TributaryConnectionTable = TributaryConnectionTable,
  CASMNodes = data.frame(NodeName = c("test", "Manawatu at Teachers College"), nzsgmnt
    = c(7140020, 7239110))
)
}
\arguments{
\item{CASMRECNetwork}{An REC V2 network (either dataframe of spatial dataframe), with at least nzsgmnt, CASMTrib and CASMTrib_Loc attributes}

\item{CASMNodes}{A dataframe of node names and REC V2 reach number (i.e. nzsgmnt attribute) of the nodes for which the CASM table is to be prepared}
}
\value{
A dataframe of CASM Node names, CASM tributary, CASM tributary location
}
\description{
This function generates a data frame of CASM Node names, CASM tributary, CASM tributary location
}
\author{
Tim Kerr, \email{Tim.Kerr@Rainfall.NZ}
}
\keyword{CASM}
\keyword{Classification}
\keyword{Environment}
\keyword{REC}
\keyword{River}
