% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASM_PrePostProcessorFunctions.R
\name{DiffuseLoadTableCreator}
\alias{DiffuseLoadTableCreator}
\title{A function to  associate the leach rate with each zone-landuse-LUC combination and find the average leach rate and area for each combination.}
\usage{
DiffuseLoadTableCreator(
  ZoneLanduseLUCRaster = ZoneLanduseLUCRaster,
  LeachRates = LeachRateRaster
)
}
\arguments{
\item{ZoneLanduseLUCRaster}{A raster of the water management zone - land use - LUC intersection.}

\item{LeachRates}{A raster of leach rates. The output from the LeachRateRasterCreator function.}
}
\value{
A data frame of average leach rates for each water management zone-landuse-LUC combination
}
\description{
A function to  associate the leach rate with each zone-landuse-LUC combination and find the average leach rate and area for each combination.
}
\author{
Tim Kerr, \email{Tim.Kerr@Rainfall.NZ}
}
\keyword{CASM}
\keyword{Classification}
\keyword{Environment}
\keyword{REC}
\keyword{River}
