% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASM_PrePostProcessorFunctions.R
\name{LeachRateRasterCreator}
\alias{LeachRateRasterCreator}
\title{A function to combine spatial data sources of climate, land use, irrigation and soil properties, and uses them to lookup leach rates based on a Ministry of Primary Industries designation. The processing is carried out at a 250 m x 250 m grid scale.}
\usage{
LeachRateRasterCreator(
  ClimateData = file.path(GISDataDirectory, ClimateShapeFileName),
  LanduseData = file.path(GISDataDirectory, SubZoneLanduseLUCShapeFileName),
  IrrigatedLandData = file.path(GISDataDirectory, IrrigatedLandShapeFileName),
  IrrigableLandData = file.path(GISDataDirectory, IrrigableLandShapeFileName),
  PAWData = file.path(GISDataDirectory, PAWShapeFileName),
  MPILeachRateData = file.path(DataDirectory, MPILeachRateLUTFile),
  LanduseToLanduseLUT = file.path(DataDirectory, LanduseToLanduseLUTFile)
)
}
\arguments{
\item{ClimateData}{A spatial data file (in ESRI polygon shapefile format) of the climate regions, as used by the MPI lookup table. Unknown origins.}

\item{LanduseData}{A spatial data file (in ESRI polygon shapefile format) of the land use, as used by Horizons. Unknown origins.}

\item{IrrigatedLandData}{A spatial data file (in ESRI polygon shapefile format) of the irrigated land, as used by the MPI lookup table. A "1" represents irrigated. Originally sourced from the  \href{https://data.mfe.govt.nz/layer/90838-irrigated-land-area-2017/}{MFE data portal}}

\item{IrrigableLandData}{A spatial data file (in ESRI polygon shapefile format) of land that is irrigable, as used by the MPI lookup table. A "1" represents irrigable. This is based on slope and elevation, but the exact source and details of preparation are unknown.}

\item{PAWData}{A spatial data file (in ESRI polygon shapefile format) of the plant available water (also known as profile available water), as used by the MPI lookup table. Originally sourced from the Fundamental Soils Layer data available from the \href{https://lris.scinfo.org.nz/layer/48100-fsl-profile-available-water/}{Landcare - Manaaki Whenua data portal}}

\item{MPILeachRateData}{A csv file with the first 5 columns (names are not important) providing, in order: "Climate";"Landuse";"Plant Available Water depth";"Irrigable Land";"Irrigated Land". A  sixth column labelled "N_loss" must also be present.}

\item{LanduseToLanduseLUT}{A csv file with two columns, one designated as "MPIClassNo" with the Ministry for Primary Industries land use classes, as used in the MPILeachRateDatalookup table. The other column designated "Horizons_Landuse_Name" with the land use names as provided in the LanduseData spatial data file.}
}
\value{
A raster object of leach rates
}
\description{
This function generates a raster object of leach rates
}
\author{
Tim Kerr, \email{Tim.Kerr@Rainfall.NZ}
}
\keyword{CASM}
\keyword{Classification}
\keyword{Environment}
\keyword{REC}
\keyword{River}
