% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReachJunctionFinderByOrder.R
\name{ReachJunctionFinderByOrder}
\alias{ReachJunctionFinderByOrder}
\title{Find confluence reaches over a strahler order}
\usage{
ReachJunctionFinderByOrder(RECNetwork = RECReachNetwork, OrderThreshold = 7)
}
\arguments{
\item{RECNetwork}{An REC V2 network, with at least a "reach identifier"nzsegment", "TO_NODE", "FROM_NODE" and "Stream_Orde" attributes.}

\item{OrderThreshold}{The strahler order which to aggregate the watersheds to}
}
\value{
A vector of reach ID's (nzsegment numbers)
}
\description{
This function finds the reaches immediately above confluences, that have a strahler order above a threshold
}
\author{
Tim Kerr, \email{Tim.Kerr@Rainfall.NZ}
}
\keyword{Classification}
\keyword{Environment}
\keyword{REC}
\keyword{River}
