% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASM_PrePostProcessorFunctions.R
\name{ReachLabeler}
\alias{ReachLabeler}
\title{A function to add labels to the REC network, based on the result of the network labeler and the OutletReachname look up table}
\usage{
ReachLabeler(
  NetworkLabels = NetworkLabelList,
  OutletReachNamesLookUpTable = OutletReachNames
)
}
\arguments{
\item{NetworkLabels}{Ideally the output from the NetworkLabeler function. A list of dataframes (one for each independent river network) with the REC V2 "nzsgmnt" of confluence reaches and "Label" which is the assigned tributary number that the confluence is the start of.}

\item{OutletReachNamesLookUpTable}{A dataframe of names and REC V2 reach number (i.e. nzsgmnt attribute) for the lowest reach in each independent catchment.}
}
\value{
A dataframe of reach numbers ("nzsgmnt") and tributary labels ("Label")
}
\description{
This function generates a data frame of CASM Node names, CASM tributary, CASM tributary location
}
\author{
Tim Kerr, \email{Tim.Kerr@Rainfall.NZ}
}
\keyword{CASM}
\keyword{Classification}
\keyword{Environment}
\keyword{REC}
\keyword{River}
