% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASM_PrePostProcessorFunctions.R
\name{TributaryConnectionCreator}
\alias{TributaryConnectionCreator}
\title{A function to determine where each tributary connects to its parent river}
\usage{
TributaryConnectionCreator(
  RECNetwork = CompleteSpatialNetwork,
  TributaryLabelList = TribLabelList,
  OutletReachNameLookUpTable = OutletReachNames
)
}
\arguments{
\item{RECNetwork}{An REC V2 network (either dataframe of spatial dataframe)}

\item{TributaryLabelList}{A list of tributary labels. One llist for each independent/disconnected catchment. The output of the NetworkLabeler() function}

\item{OutletReachNameLookUpTable}{A data frame of outlet reach nzsgmnt numbers and real-world names.}
}
\value{
A dataframe. One row for each tributary. Columns of nzsgmnt, tributary name, confluence name, confluence location in kilometres.
}
\description{
This function provides the river network connectivity table used by CASM.
CASM understands a network in terms of Tributary Names, Confluence Names and Confluence Locations.
Confluence Name is the name of the river that a tributary flows into.
Confluence Location is the distance along the confluence from its start, that the tributary joins it.
}
\author{
Tim Kerr, \email{Tim.Kerr@Rainfall.NZ}
}
\keyword{CASM}
\keyword{Classification}
\keyword{Environment}
\keyword{REC}
\keyword{River}
