% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASM_PrePostProcessorFunctions.R
\name{ZoneLanduseLUCRasterCreator}
\alias{ZoneLanduseLUCRasterCreator}
\title{A function to  rasterize the zone-landuse-LUC data.}
\usage{
ZoneLanduseLUCRasterCreator(
  ZoneLanduseLUCPolygons = SubZoneLanduseLUCSpatial,
  LeachRates = LeachRateRaster
)
}
\arguments{
\item{ZoneLanduseLUCPolygons}{A polygon spatial data file of the water management zone - land use - LUC intersection.}

\item{LeachRates}{A raster of leach rates. The output from the LeachRateRasterCreator function. Used to allign the output raster}
}
\value{
A raster of the zone-landuse-LUC combination
}
\description{
This function generates a raster object of the zone-landuse-LUC data, writes a copy to file as a tiff, and writes a csv file of a look up table of the raster values to the combined class names
}
\author{
Tim Kerr, \email{Tim.Kerr@Rainfall.NZ}
}
\keyword{CASM}
\keyword{Classification}
\keyword{Environment}
\keyword{REC}
\keyword{River}
